Allow subscriber subset to accept query ids
