+++ flowmachine/flowmachine/core/server/query_schemas/displacement.py
@@ -7,7 +7,8 @@ from marshmallow.validate import OneOf
 from marshmallow_oneofschema import OneOfSchema
 
 from flowmachine.features import Displacement
-from .custom_fields import EventTypes, SubscriberSubset, Statistic, ISODateTime
+from .custom_fields import EventTypes, Statistic, ISODateTime
+from .subscriber_subset import SubscriberSubset
 from .daily_location import DailyLocationSchema
 from .modal_location import ModalLocationSchema
 from .base_query_with_sampling import (
