+++ flowmachine/flowmachine/core/server/query_schemas/topup_balance.py
@@ -6,7 +6,8 @@ from marshmallow import fields
 from marshmallow.validate import OneOf
 
 from flowmachine.features import TopUpBalance
-from .custom_fields import Statistic, SubscriberSubset, ISODateTime
+from .custom_fields import Statistic, ISODateTime
+from .subscriber_subset import SubscriberSubset
 from .base_query_with_sampling import (
     BaseQueryWithSamplingSchema,
     BaseExposedQueryWithSampling,
