+++ flowmachine/flowmachine/core/server/query_schemas/unique_location_counts.py
@@ -8,7 +8,8 @@ from marshmallow.validate import OneOf
 from flowmachine.features import UniqueLocationCounts
 from . import BaseExposedQuery
 from .base_schema import BaseSchema
-from .custom_fields import EventTypes, SubscriberSubset, ISODateTime
+from .custom_fields import EventTypes, ISODateTime
+from .subscriber_subset import SubscriberSubset
 from .aggregation_unit import AggregationUnitMixin
 
 __all__ = ["UniqueLocationCountsSchema", "UniqueLocationCountsExposed"]
